Mod for internal style
